# specs/000-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Manager API
  version: 0.3.0
  description: Contrato inicial do MVP. Endpoints autenticados usam sessão em cookie.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
...
paths:
...
  /operators:
    get:
      tags: [Operators]
      operationId: listOperators
      summary: Listar Operators do tenant corrente
      description: Requer papel TenantOwner; tenant é derivado da sessão.
      x-required-role: TenantOwner
      x-tenant-scoped: true
      x-requirements: [US-009, FR-002, FR-026, BR-012]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Página de memberships Operator do tenant corrente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorPage' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /operators/{membershipId}/deactivate:
    post:
      tags: [Operators]
      operationId: deactivateOperator
      summary: Desativar Operator e invalidar sessões
      description: Requer papel TenantOwner e membership Operator do tenant corrente.
      x-required-role: TenantOwner
      x-tenant-scoped: true
      x-requirements: [US-009, FR-002, FR-026, FR-028, BR-012, BR-015]
      parameters:
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/AntiforgeryToken'
      responses:
        '200':
          description: Operator desativado e sessões anteriores invalidadas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorSummary' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /operators/{membershipId}/reactivate:
    post:
      tags: [Operators]
      operationId: reactivateOperator
      summary: Reativar Operator
      description: Requer papel TenantOwner e membership Operator do tenant corrente; cookies antigos permanecem inválidos.
      x-required-role: TenantOwner
      x-tenant-scoped: true
      x-requirements: [US-009, FR-002, FR-026, FR-028, BR-012, BR-015]
      parameters:
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/AntiforgeryToken'
      responses:
        '200':
          description: Operator reativado; novo login é necessário
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorSummary' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-waai-session
  parameters:
    AntiforgeryToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema: { type: string, minLength: 16 }
    IfMatch:
      in: header
      name: If-Match
      required: true
      schema: { type: string }
...
    MembershipId:
      in: path
      name: membershipId
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    PageLimit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    OperatorSummary:
      type: object
      required: [membershipId, userId, email, status, version]
      properties:
        membershipId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        status: { type: string, enum: [Invited, Active, Disabled] }
        version: { type: integer, minimum: 0 }
        invitationExpiresAt: { type: [string, 'null'], format: date-time }
    OperatorPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/OperatorSummary' }
        nextCursor: { type: [string, 'null'] }
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
security:
  - sessionCookie: []